backend/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
chromadb>=0.6.0
python-dotenv>=1.0.1
pydantic>=2.10.3
python-multipart>=0.0.20
aiofiles>=24.1.0
httpx>=0.28.1
anthropic>=0.44.0
